Revert "chore: setup app for e2e and dev"
This reverts commit 55b2057c5bbeb4753275684b8d1da4e89f77b55b.
