--- dev/dev.tsx
@@ -1,2 +1,0 @@
-import { initApp } from './initApp';
-initApp();
