--- dev/e2e.tsx
@@ -1,3 +1,0 @@
-import { initApp } from './initApp';
-
-(window as any).initApp = initApp;
